Modified Database & Website Layout
Included a String data type called "eventBlocks" that would be used to hold the String passed from the front end storing what blocks of time the user of our app has created and saved from the calendar.  NOTE: All I have implemented is the attribute to be used for storing the time blocks, I have NOT implemented the necessary back end operations to store what is obtained from the front end into the database.

Additionally: I modified the layout of our website.  I believe that the developer section is better to be in the about us section, as it is more relevant there.  Other minor changes were made too.